Refactor: consolidate database types into single types file
