--- lib/supabase/database.types.ts
@@ -1,258 +1,0 @@
-/**
- * Database types for Supabase
- * Generated types for all tables - update using: supabase gen types typescript
- */
-
-export type BoxType = 
-  | 'monthly-standard' 
-  | 'monthly-premium' 
-  | 'monthly-premium-monthly' 
-  | 'monthly-premium-seasonal' 
-  | 'onetime-standard' 
-  | 'onetime-premium';
-
-// ============================================================================
-// Box Types Table
-// ============================================================================
-
-export interface BoxTypeRow {
-  id: string;
-  name: string;
-  description: string | null;
-  price_eur: number;
-  is_subscription: boolean;
-  is_premium: boolean;
-  frequency: string | null;
-  sort_order: number;
-  is_enabled: boolean;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface BoxTypeInsert {
-  id: string;
-  name: string;
-  description?: string | null;
-  price_eur: number;
-  is_subscription?: boolean;
-  is_premium?: boolean;
-  frequency?: string | null;
-  sort_order?: number;
-  is_enabled?: boolean;
-}
-
-export interface BoxTypeUpdate {
-  id?: string;
-  name?: string;
-  description?: string | null;
-  price_eur?: number;
-  is_subscription?: boolean;
-  is_premium?: boolean;
-  frequency?: string | null;
-  sort_order?: number;
-  is_enabled?: boolean;
-}
-
-// ============================================================================
-// Promo Codes Table
-// ============================================================================
-
-export interface PromoCodeRow {
-  id: string;
-  code: string;
-  discount_percent: number;
-  description: string | null;
-  is_enabled: boolean;
-  starts_at: string | null;
-  ends_at: string | null;
-  max_uses: number | null;
-  current_uses: number;
-  min_order_value_eur: number | null;
-  applicable_box_types: string[] | null;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface PromoCodeInsert {
-  code: string;
-  discount_percent: number;
-  description?: string | null;
-  is_enabled?: boolean;
-  starts_at?: string | null;
-  ends_at?: string | null;
-  max_uses?: number | null;
-  current_uses?: number;
-  min_order_value_eur?: number | null;
-  applicable_box_types?: string[] | null;
-}
-
-export interface PromoCodeUpdate {
-  code?: string;
-  discount_percent?: number;
-  description?: string | null;
-  is_enabled?: boolean;
-  starts_at?: string | null;
-  ends_at?: string | null;
-  max_uses?: number | null;
-  current_uses?: number;
-  min_order_value_eur?: number | null;
-  applicable_box_types?: string[] | null;
-}
-
-// ============================================================================
-// Options Table
-// ============================================================================
-
-export type OptionSetId = 'sports' | 'colors' | 'flavors' | 'dietary' | 'sizes';
-
-export interface OptionRow {
-  id: string;
-  option_set_id: string;
-  label: string;
-  value: string | null;
-  sort_order: number;
-  is_enabled: boolean;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface OptionInsert {
-  id: string;
-  option_set_id: string;
-  label: string;
-  value?: string | null;
-  sort_order?: number;
-  is_enabled?: boolean;
-}
-
-export interface OptionUpdate {
-  id?: string;
-  option_set_id?: string;
-  label?: string;
-  value?: string | null;
-  sort_order?: number;
-  is_enabled?: boolean;
-}
-
-// ============================================================================
-// Site Config Table
-// ============================================================================
-
-export interface SiteConfigRow {
-  key: string;
-  value: string;
-  description: string | null;
-  value_type: string;
-  created_at: string;
-  updated_at: string;
-}
-
-export interface SiteConfigInsert {
-  key: string;
-  value: string;
-  description?: string | null;
-  value_type?: string;
-}
-
-export interface SiteConfigUpdate {
-  key?: string;
-  value?: string;
-  description?: string | null;
-  value_type?: string;
-}
-
-// ============================================================================
-// Preorders Table (existing)
-// ============================================================================
-
-export interface PreorderInsert {
-  full_name: string;
-  email: string;
-  phone?: string | null;
-  box_type: BoxType;
-  wants_personalization: boolean;
-  sports?: string[] | null;
-  sport_other?: string | null;
-  colors?: string[] | null;
-  flavors?: string[] | null;
-  flavor_other?: string | null;
-  size_upper?: string | null;
-  size_lower?: string | null;
-  dietary?: string[] | null;
-  dietary_other?: string | null;
-  additional_notes?: string | null;
-  // Promo code fields
-  promo_code?: string | null;
-  discount_percent?: number | null;
-  original_price_eur?: number | null;
-  final_price_eur?: number | null;
-}
-
-export interface Preorder extends PreorderInsert {
-  id: string;
-  order_id: string;
-  created_at: string;
-  updated_at: string;
-}
-
-// ============================================================================
-// Database Schema Type
-// ============================================================================
-
-export interface Database {
-  public: {
-    Tables: {
-      preorders: {
-        Row: Preorder;
-        Insert: PreorderInsert;
-        Update: Partial<PreorderInsert>;
-        Relationships: [];
-      };
-      box_types: {
-        Row: BoxTypeRow;
-        Insert: BoxTypeInsert;
-        Update: BoxTypeUpdate;
-        Relationships: [];
-      };
-      promo_codes: {
-        Row: PromoCodeRow;
-        Insert: PromoCodeInsert;
-        Update: PromoCodeUpdate;
-        Relationships: [];
-      };
-      options: {
-        Row: OptionRow;
-        Insert: OptionInsert;
-        Update: OptionUpdate;
-        Relationships: [];
-      };
-      site_config: {
-        Row: SiteConfigRow;
-        Insert: SiteConfigInsert;
-        Update: SiteConfigUpdate;
-        Relationships: [];
-      };
-    };
-    Views: Record<string, never>;
-    Functions: {
-      calculate_box_prices: {
-        Args: { p_promo_code: string | null };
-        Returns: {
-          box_type_id: string;
-          box_type_name: string;
-          original_price_eur: number;
-          original_price_bgn: number;
-          discount_percent: number;
-          discount_amount_eur: number;
-          discount_amount_bgn: number;
-          final_price_eur: number;
-          final_price_bgn: number;
-        }[];
-      };
-    };
-    Enums: {
-      box_type: BoxType;
-    };
-    CompositeTypes: Record<string, never>;
-  };
-}
